backend/requirements.txt
fastapi==0.115.12
uvicorn[standard]==0.34.2
python-multipart==0.0.20
rembg[cpu]==2.0.74
psd-tools==1.10.7
Pillow==12.1.0
PyMuPDF==1.25.3
python-pptx==1.0.2
